# swagger.yaml
openapi: 3.0.0
info:
  title: 1. Version 1
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
...
paths:
  /import:
    get:
      tags:
        - Import product
      summary: get all Import request join
      parameters:
        - name: status
          in: query
          schema:
            type: string
          example: COMPLETED
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - 'Replicate: Flow nhập hàng'
      summary: '1. ImpP:Create: Tạo yêu cầu nhập hàng'
      description: |-
        ``` json
        {
            // Định danh sản phẩm
            "product_id": 2,
            "subproduct_id": 4,
            // cập nhật lại ở step 3
            "supplier_id": null,
            "price_quotation_id": null,
            "total_cost": 0,
            // cố định
            "status": "REQUEST",
            //
            "quantity": 1000,
            "note": "abc",
            "created_by": "",
            "updated_by": ""
        }

        ```
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                product_id: 2
                subproduct_id: 4
                supplier_id: null
                price_quotation_id: null
                total_cost: 0
                status: REQUEST
                quantity: 1000
                note: abc
                created_by: ''
                updated_by: ''
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...